Phase 6b correction slice: split frontend pages, LLD edge repairs
Applies the 6b review (request-changes on d8c4e1a3):

- Frontend page physically SPLIT into FrontendBufferPage (raw + mel
  tails, fp32) and FrontendCounterPage (eight int64 counters): the
  pinned runner constructs typed views with dtype-aligned offsets and
  strides (gpu_model_runner.py:7160 @ v0.24.0/ee0da84), and a mixed
  page put the counters at byte 12,420 with a 12,484-byte stride,
  neither 8-byte aligned. Two homogeneous pages keep the exact
  12,484-byte frontend inventory and 6,314,864-byte session total.
  New source evidence reversing the earlier mixed-dtype-split
  rejection; ledger note follows in the notes repo.
- frontend_state registered in TENSOR_CLASSES; BF16_COMPUTE pins it
  fp32 explicitly (its identifier restamps by content hash; only
  FP32_BRINGUP's id is pinned by value and its mapping is unchanged).
- Final residual rule per the LLD: a final tail commits its remainder
  only when at least eight new mel frames lie past the committed
  boundary; a shorter remainder is dropped, finalization still marks
  atomically. The exact-cadence case (final - B_k = 7, dropped) and
  the zero-sample final are now test-pinned; whole-signal equality
  tests keep full length only where the final packet's residual is
  large.
- advance_session repairs: device-correct frontend allocation
  (CUDA-safe gathers), _StackedRows facade so stream_step's
  caches.channel.shape / slice-assign surface works over the gathered
  batch, the CONFIGURED two-output pre-encode drop with the full
  nine-slot mel-tail prefix (never a dynamic ceil), per-prompt LID
  grouping (mixed-language rows never share a multi-hot prompt),
  zero-new-frame rows skip encoder/LID/decode entirely, and
  session-first/continuing rows may not share a bucket.
- Stale tests reconciled: 7-slot int32 book + whole-page park
  persistence (test_state_layers), int32 window valid slot
  (test_window_state_page), 52 prefixes incl. the frontend pair
  (test_bringup_config_weights), persist-by-kind constructs the split
  pages (test_advance).

Still owed before advance_model_rows (per the review's order):
length-aware tensorized eager path (fixed-trip masked label loop,
GPU-resident emission state, padded variable-length final batches)
and CUDA differential coverage across the five geometries.

+++ tests/model_executor/models/nemotron_asr/test_window_state_page.py
@@ -79,9 +79,9 @@ WINDOW_PAGE_BYTES = (WINDOW * D_MODEL + 1) * 4
 def test_window_page_shapes_and_policy_dtypes():
     page = make_window_page()
     assert tuple(page.get_state_shape()) == ((WINDOW, D_MODEL), (1,))
-    # attention_cache + queue_state classes, both fp32 in the bring-up
+    # attention_cache (policy fp32) + the int32 valid-length control
     # policy (PORT-PREC-001/005).
-    assert page.get_state_dtype() == (torch.float32, torch.float32)
+    assert page.get_state_dtype() == (torch.float32, torch.int32)
 
 
 def test_window_page_is_now_the_largest_kind():
